Led control of the cube from App
Now I have a boolean that works as an "app-permission". When the app sends a signal, this puts the boolean to the opposite value. The whole led control is put inside an if, depending on this value.
Also, there is a communication esp -> app via sending to the app state of each side boolean (led_up and so on) each ~4 seconds.